backend/requirements.txt
arrow
attrs
autopep8 
Babel
beautifulsoup4 
bokeh 
boto3 
botocore
decorator 
Django==4.1.5
django-cors-headers==3.13.0
django-crispy-forms==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docutils 
flake8 
Flask
glob2 
h5py
Jinja2 
jinja2-time 
json5 
jsonschema 
openpyxl==3.0.10
pep8==1.7.1
pika==1.3.1
Pillow==9.2.0
PyDispatcher==2.0.5
pydocstyle
PyJWT 
pylint
pyodbc 
pyOpenSSL 
pyparsing 
pyrsistent 
PySocks
python-dateutil
PyYAML==6.0
queuelib==1.5.0
regex 
requests 
requests-file 
rope 
s3transfer
six
SQLAlchemy 
sqlparse==0.4.3
statsmodels
tabulate 
threadpoolctl
urllib3
uwsgi
w3lib
wheel
xlrd 
XlsxWriter
zipp
